Made bar lengths indicate title length.
On a disk with multiple titles, the bars are sized realtive to the
length of longest title. This makes it easier to notice short titles
(often an indication of logo "bumps" or movie trailers).

Also, the chapter's section of the bar will now always be at least one
unit long, so that the position of very short chapters is apparent.