Add --mount-timeout flag
This makes dvdrip wait a bit when asked to operate on a mountpoint. This
is handy when you want to pop in a disc and tell dvdrip to scan it as it
sometimes takes a few seconds for the disc to actually mount. This flag
defaults to a few seconds, set it to 0 to disable this feature.